Move scalar-4x64 implementation to a separate module
