+++ k256/Cargo.toml
@@ -29,9 +29,10 @@ num-bigint = "0.3"
 num-traits = "0.2"
 
 [features]
-default = ["arithmetic", "std", "field-5x52"]
+default = ["arithmetic", "std", "field-5x52", "scalar-4x64"]
 arithmetic = []
 field-5x52 = []
+scalar-4x64 = []
 nightly-bench = []
 rand = ["elliptic-curve/rand_core"]
 test-vectors = []
